fix(remarkable): require space after bujo symbol to avoid false positives
Words starting with 'a', 'o', or 'x' (e.g., "adoption", "over", "x-ray")
were incorrectly treated as having a bujo prefix, skipping the `- ` prepend.
Now requires a space after the symbol character to match actual bujo format.

+++ internal/adapter/remarkable/rmrender.go
@@ -16,7 +16,7 @@ func detectXOffset(strokes []rmStroke) float64 {
 		return halfWidth
 	}
 
-	minX, maxX := float64(math.MaxFloat64), float64(-math.MaxFloat64)
+	minX, maxX := math.MaxFloat64, -math.MaxFloat64
 	for _, s := range strokes {
 		for _, p := range s.Points {
 			x := float64(p.X)
